+++ src/editor/DeconzEditor.js
@@ -182,9 +182,9 @@ class DeconzEditor {
     getI18n(prefix, suffix, data = {}) {
         let _path = prefix;
         if (suffix) _path += `.${suffix}`;
-        data.defaultValue = '__undefined__';
+        data.defaultValue = '_deconz_undefined_';
         const value = RED._(_path, data);
-        if (value === '__undefined__') return;
+        if (value === '_deconz_undefined_') return;
         return value;
     }
 
@@ -425,6 +425,51 @@ class DeconzEditor {
 
     //#endregion
 
+    // From https://stackoverflow.com/questions/6832596
+    static versionCompare(v1, v2, options = {}) {
+        const lexicographical = options && options.lexicographical;
+        const zeroExtend = options && options.zeroExtend;
+        let v1parts = v1.split('.');
+        let v2parts = v2.split('.');
+
+        function isValidPart(x) {
+            return (lexicographical ? /^\d+[A-Za-z]*$/ : /^\d+$/).test(x);
+        }
+
+        if (!v1parts.every(isValidPart) || !v2parts.every(isValidPart)) {
+            return NaN;
+        }
+
+        if (zeroExtend) {
+            while (v1parts.length < v2parts.length) v1parts.push("0");
+            while (v2parts.length < v1parts.length) v2parts.push("0");
+        }
+
+        if (!lexicographical) {
+            v1parts = v1parts.map(Number);
+            v2parts = v2parts.map(Number);
+        }
+
+        for (let i = 0; i < v1parts.length; ++i) {
+            if (v2parts.length === i) {
+                return 1;
+            }
+
+            if (v1parts[i] === v2parts[i]) {
+                //continue;
+            } else if (v1parts[i] > v2parts[i]) {
+                return 1;
+            } else {
+                return -1;
+            }
+        }
+
+        if (v1parts.length !== v2parts.length) {
+            return -1;
+        }
+
+        return 0;
+    }
 
 }